airport-drf/airport_config/urls.py
"""
URL configuration for airport_config project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, include
from rest_framework import routers

from airport import views

router = routers.DefaultRouter()
router.register(r'airports', views.AirportViewSet)
router.register(r'routes', views.RouteViewSet)
router.register(r'airplane-types', views.AirplaneTypeViewSet)
router.register(r'airplanes', views.AirplaneViewSet)
router.register(r'flights', views.FlightViewSet)
router.register(r'crew', views.CrewViewSet)
router.register(r'orders', views.OrderViewSet)
router.register(r'tickets', views.TicketViewSet)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include(router.urls)),
]

