airport-drf/requirements.txt
asgiref==3.9.1
attrs==25.3.0
colorama==0.4.6
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
psycopg==3.2.10
psycopg-binary==3.2.10
Pygments==2.19.2
PyJWT==2.10.1
pyrsistent==0.20.0
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
wheel==0.45.1
zipp==3.23.0